feat: refactor toast functionality to use new ``toast`` functions and ``Toaster`` component
